Unzip all the time if needed
